Add "--noshutdown" option to loader
Added a "--noshutdown" option to loader which prevents a shutdown
(and a poweroff), after main() returns.

Note that this doesn't just replace poweroff() by halt() - with this
option, the system *isn't* halted after main() returns, but rather
continues as usual - possibly still running other threads that main()
didn't wait for, running various system threads and services, and so on.