Fix errors in serial console setup
Fix two errors in IsaSerialConsole::reset(). The visible effect of these
errors was a spurious \001 (^A) character which always appeared at the
beginning of every OSV's run output.

The two errors were:
1. The DLAB bit on the LCR was incorrectly set (a spurious
   "& LCR_DIVISOR_LATCH_ACCESS_BIT" appeared in the code), causing the
   next outb(1) to go to the output register (and output a "1" character)
   instead of setting the baud rate.
2. When writing the FCR register, the "ioport + " base forgotten.

I took the opportunity and added better names for the constants,
removed a couple of duplicate constants, and added comments.