kvmclock: don't wastefully compute the wall time at every clock measure
I am pretty sure kvmclock only sets the wall time when we write to the MSR,
so there is no point in keep calculating it every time we call system_time().

I am not changing it to a const function because we *may* have to call it
more than once in the future, if we ever suspend, or the likes. But for now,
using a pre-computed value is enough.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>