Make lock-free mutex our default mutex
This patch turns on the flag which switches all our code to use the
lock-free mutex instead of the spinlock-based mutex.

It's time we start using the lock-free mutex, which is stable enough by
now - but please let me know if you do experience any performance problem,
or bugs, related to the new mutex.

If you need to disable the new mutex temporarily and return to the old,
just change the "#define LOCKFREE_MUTEX" in osv/mutex.h to #undef.