Fix "ifconfig" corrupting accept_mtx
ifconfig used to call sofree(), which assumed accept_mtx was taken, which
wasn't true, resulting in either an assertion failure (if we implement
assert_mtx - see next patch) or a mutex corruption (if assert_mtx does
nothing).

Instead, we should call soclose(). This wasn't very hard to figure out,
given the comment in socreate() saying "The socket should be closed with
soclose()." :-)